requirements-test.txt
pytest==6.2.3
pytest-aiohttp==0.3.0
pytest-asyncio==0.15.1
pytest-benchmark==3.4.1